fix/feat: thread safety, game logic fixes, and Avalon UX improvements
  - Migrate room managers to ConcurrentDictionary with atomic APIs
  - Fix settlement race condition with Interlocked compare-and-swap
  - Fix disposed scoped service in 30-min disconnect Task.Run callback
  - Fix host reassignment on player disconnect
  - Fix TurnTimerService missing balance data in DTOs and broadcasts
  - Fix RebuildRoleConfig crash when evil roles exceed player count
  - Fix BlackJack player leave during betting not triggering AllBetsPlaced
  - Fix PlaceBet allowing forfeited players to bet
  - Avalon: add vote/mission progress visibility (who acted vs waiting)
  - Avalon: add ready status display in post-game settlement screen
  - Avalon: require full room (playerCount == maxPlayers) to start game
  - Avalon: prevent assassin from targeting evil players (including Oberon)
  - Avalon: show "Ally" tag on evil teammates during assassination phase
  - Avalon: exclude Percival from shield bonus (+0.5) on assassination
  - Add coverlet.runsettings to exclude Migrations from coverage
  - Fix flaky tests using deterministic deck and correct assertions

+++ BoardGames.Tests/Controllers/AuthControllerTests.cs
@@ -162,8 +162,8 @@ public class AuthControllerTests : IDisposable
         var result = await _controller.GetBalances();
 
         var okResult = Assert.IsType<OkObjectResult>(result);
-        var balances = Assert.IsType<Dictionary<string, int>>(okResult.Value);
-        Assert.Equal(500, balances["BlackJack"]);
-        Assert.Equal(1000, balances["Avalon"]);
+        var balances = Assert.IsType<Dictionary<string, decimal>>(okResult.Value);
+        Assert.Equal(500m, balances["BlackJack"]);
+        Assert.Equal(1000m, balances["Avalon"]);
     }
 }
