+++ BoardGames/Data/IGameBalanceRepository.cs
@@ -5,6 +5,6 @@ namespace BoardGames.Data;
 public interface IGameBalanceRepository
 {
     Task<GameBalance> GetOrCreate(int userId, GameType gameType);
-    Task UpdateBalance(int userId, GameType gameType, int delta);
-    Task<List<(string Nickname, int Balance)>> GetTopBalances(GameType gameType, int count);
+    Task UpdateBalance(int userId, GameType gameType, decimal delta);
+    Task<List<(string Nickname, decimal Balance)>> GetTopBalances(GameType gameType, int count);
 }
