+++ BoardGames/Models/GameBalance.cs
@@ -11,5 +11,5 @@ public class GameBalance
     public int Id { get; set; }
     public int UserId { get; set; }
     public GameType GameType { get; set; }
-    public int Balance { get; set; } = 1000;
+    public decimal Balance { get; set; } = 1000;
 }
